Requirement 4 – the game's output A few outputs should be added to let the players know the current status of the game.
We want feedback when either an event or an error occurs within the game. The output shows the status of the board after every move.
No output channel is specified. To make it easier, we decided to use the system standard output to print an event when it occurs. A few lines have been added on every action to let the user know about the status of the game