Requirement 8 – win condition (IV) The last requirement is the last win condition. It is pretty similar to the last two; in this case, in a diagonal direction.
If a player inserts a disc and connects more than three discs of his color in a straight diagonal line, then that player wins.
This is a possible implementation for this last requirement. The code is very similar to the other win conditions because the same statement must be fulfilled.

What we have got is a class with one constructor, three public methods, and three private methods. The logic of the application is distributed among all methods. The biggest flaw here is that this class is very difficult to maintain. The crucial methods, such as checkWinCondition, are non-trivial, with potential for bug entries in future modifications.

If you want to take a look at the full code, you can find it in the https://bitbucket.org/vfarcic/tdd-java-ch05-design.git repository.

We made this small example to demonstrate the common problems with this approach. Topics such as the SOLID principle requires a bigger project to become more illustrative.

In large projects with hundreds of classes, the problems become hours wasted in a sort of surgical development. Developers spend a lot of their time investigating tricky code and understanding how it works, instead of creating new features.