Requirement 6 – win condition (II) This is the first win condition requirement for players.
If a player inserts a disc and connects more than three discs of his color in a straight vertical line, then that player wins.
In fact, this requires one single check. If the current inserted disc connects other three discs in a vertical line, the current player wins the game

There are a couple of changes to the putDiscInColumn method. Also, a new method called checkWinner has been created