Requirement 8 – win condition (IV) The last requirement is the last win condition.
If a player inserts a disc and connects more than three discs of his color in a straight diagonal line, then that player wins.
We need to perform valid game movements to achieve the condition. In this case, we need to test both diagonals across the board: from top-right to bottom-left and from bottom-right to top-left. The following tests use a list of columns to recreate a full game to reproduce the scenario under test

Again, the checkWinner method needs to be modified, adding new board verifications